refactor(): image gallery, add placeholder
